network restructure

+++ network-impl/src/main/java/io/nuls/network/service/impl/netty/ServerChannelHandler.java
@@ -5,17 +5,20 @@ import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.channel.socket.SocketChannel;
+import io.nuls.core.chain.entity.Block;
 import io.nuls.core.context.NulsContext;
 import io.nuls.core.utils.log.Log;
 import io.nuls.core.utils.network.IpUtil;
 import io.nuls.core.utils.spring.lite.annotation.Autowired;
 import io.nuls.network.constant.NetworkConstant;
 import io.nuls.network.entity.Node;
-import io.nuls.network.entity.NodeGroup;
+import io.nuls.network.message.entity.GetVersionEvent;
+import io.nuls.network.message.entity.HandshakeEvent;
 import io.nuls.network.service.NetworkService;
 
 import java.net.InetAddress;
 import java.nio.ByteBuffer;
+import java.util.Map;
 
 /**
  * @author Vive
@@ -30,44 +33,38 @@ public class ServerChannelHandler extends ChannelInboundHandlerAdapter {
     public void channelRegistered(ChannelHandlerContext ctx) throws Exception {
         SocketChannel channel = (SocketChannel) ctx.channel();
         String nodeId = IpUtil.getNodeId(channel.remoteAddress());
-//        Log.info("---------------------- server channelRegistered ------------------------- " + nodeId);
+        System.out.println("---------------------- server channelRegistered ------------------------- " + nodeId);
+
         String remoteIP = channel.remoteAddress().getHostString();
-
-        NodeGroup outGroup = networkService.getNodeGroup(NetworkConstant.NETWORK_NODE_OUT_GROUP);
-        for (Node node : outGroup.getNodes().values()) {
+        Map<String, Node> nodeMap = getNetworkService().getNodes();
+        for (Node node : nodeMap.values()) {
             if (node.getIp().equals(remoteIP)) {
-                String localIP = InetAddress.getLocalHost().getHostAddress();
-                boolean isLocalServer = IpUtil.judgeIsLocalServer(localIP, remoteIP);
-
-                if (!isLocalServer) {
-                    ctx.channel().close();
-                    return;
-                } else {
-                    getNetworkService().removeNode(node.getId());
+                // 当前节点存在已主动连接出去的node，则只保留一个连接
+                if (node.getType() == Node.OUT) {
+                    String localIP = InetAddress.getLocalHost().getHostAddress();
+                    boolean isLocalServer = IpUtil.judgeIsLocalServer(localIP, remoteIP);
+                    if (!isLocalServer) {
+                        // 关闭当前被动连接
+                        ctx.channel().close();
+                        return;
+                    } else {
+                        // 关闭已有主动连接
+                        getNetworkService().removeNode(node.getId());
+                    }
                 }
             }
         }
+
         // if has a node with same ip, and it's a out node, close this channel
         // if More than 10 in nodes of the same IP, close this channel
         int count = 0;
-        for (Node n : getNetworkService().getNodes().values()) {
+        for (Node n : nodeMap.values()) {
             if (n.getIp().equals(remoteIP)) {
                 count++;
-                if (count >= Node.SAME_IP_MAX_COUNT) {
+                if (count >= NetworkConstant.SAME_IP_MAX_COUNT) {
                     ctx.channel().close();
                     return;
                 }
-//
-//                if (n.getType() == Node.OUT && n.getStatus() == Node.HANDSHAKE) {
-//                    ctx.channel().close();
-//                    return;
-//                } else {
-//                    count++;
-//                    if (count == 10) {
-//                        ctx.channel().close();
-//                        return;
-//                    }
-//                }
             }
         }
     }
@@ -76,27 +73,39 @@ public class ServerChannelHandler extends ChannelInboundHandlerAdapter {
     public void channelActive(ChannelHandlerContext ctx) throws Exception {
         SocketChannel channel = (SocketChannel) ctx.channel();
         String nodeId = IpUtil.getNodeId(channel.remoteAddress());
-        Log.info("---------------------- server Active ------------------------- " + nodeId);
+        System.out.println("---------------------- server channelActive ------------------------- " + nodeId);
+
         String channelId = ctx.channel().id().asLongText();
-        //SocketChannel channel = (SocketChannel) ctx.channel();
         NioChannelMap.add(channelId, channel);
-        Node node = new Node(Node.IN, channel.remoteAddress().getHostString(), channel.remoteAddress().getPort(), channelId);
+        Node node = new Node(channel.remoteAddress().getHostString(), channel.remoteAddress().getPort(), Node.IN);
+        node.setChannelId(channelId);
         node.setStatus(Node.CONNECT);
-        getNetworkService().addNode(node);
-//        getNetworkService().addNodeToGroup(NetworkConstant.NETWORK_NODE_IN_GROUP, node);
+        boolean success = getNetworkService().addConnNode(node);
+        if (!success) {
+            ctx.channel().close();
+            return;
+        }
+        HandshakeEvent event = new HandshakeEvent(NetworkConstant.HANDSHAKE_SEVER_TYPE);
+        System.out.println("---sever send handshake---" + node.getId());
+        getNetworkService().sendToNode(event, nodeId, false);
     }
 
     @Override
     public void channelInactive(ChannelHandlerContext ctx) throws Exception {
         SocketChannel channel = (SocketChannel) ctx.channel();
+        String nodeId = IpUtil.getNodeId(channel.remoteAddress());
+        System.out.println(" ---------------------- server channelInactive ------------------------- " + nodeId);
+
         String channelId = ctx.channel().id().asLongText();
         NioChannelMap.remove(channelId);
-        String nodeId = IpUtil.getNodeId(channel.remoteAddress());
-        Log.info("---------------------- server remove ------------------------- " + nodeId);
         Node node = getNetworkService().getNode(nodeId);
         if (node != null) {
             if (channelId.equals(node.getChannelId())) {
                 getNetworkService().removeNode(nodeId);
+            } else {
+                System.out.println("--------------channel id different----------------------");
+                System.out.println(node.getChannelId());
+                System.out.println(channelId);
             }
         }
     }
@@ -111,6 +120,7 @@ public class ServerChannelHandler extends ChannelInboundHandlerAdapter {
     public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
         SocketChannel channel = (SocketChannel) ctx.channel();
         String nodeId = IpUtil.getNodeId(channel.remoteAddress());
+//        System.out.println(" ---------------------- server channelRead ------------------------- " + nodeId);
         Node node = getNetworkService().getNode(nodeId);
         if (node != null && node.isAlive()) {
             ByteBuf buf = (ByteBuf) msg;
